+++ index.html
@@ -8,7 +8,6 @@
   <link rel="shortcut icon" href="favicon.ico" type="image/x-icon">
   <title>Кекстаграм</title>
   <script src="./js/functions.js"></script>
-  <script src="./js/main.js"></script>
 </head>
 
 <body>
@@ -228,6 +227,8 @@
       </div>
     </section>
   </template>
+
+  <script src="./main.js" type="module"></script>
 
 </body>
 </html>
